requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
bcrypt==3.2.0
billiard==3.3.0.23
bleach==3.3.0
celery==3.1.26.post2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
Django==3.1.6
django-bleach==0.6.1
django-braces==1.14.0
django-celery==3.3.1
django-choices==1.7.1
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-countries==7.0
django-crispy-forms==1.11.0
django-currentuser==0.5.2
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.0
django-js-asset==1.2.2
django-loginas==0.3.9
django-nested-inline==0.4.2
django-phonenumber-field==5.0.0
django-recaptcha==2.0.6
django-redis==4.12.1
django-tabbed-admin==1.0.4
django-wkhtmltopdf==3.3.0
docopt==0.6.2
Fabric3==1.14.post1
Faker==5.8.0
idna==2.10
kombu==3.0.37
mixer==7.1.2
packaging==20.9
paramiko==2.7.2
phonenumbers==8.12.17
Pillow==8.1.0
pipreqs==0.4.11
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
text-unidecode==1.3
urllib3==1.26.3
webencodings==0.5.1
yarg==0.1.9
